+++ spring-xsuaa/src/main/java/com/sap/cloud/security/xsuaa/token/OAuth2AuthenticationConverter.java
@@ -1,18 +1,19 @@
 package com.sap.cloud.security.xsuaa.token;
 
 import com.sap.cloud.security.xsuaa.extractor.AuthoritiesExtractor;
-import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.core.OAuth2AccessToken;
-import org.springframework.security.oauth2.core.OAuth2AuthenticatedPrincipal;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.security.oauth2.server.resource.authentication.BearerTokenAuthentication;
 
-import java.util.Collection;
-import java.util.Map;
 import java.util.Objects;
 
 /**
  * @deprecated OAuth2AuthenticationConverter won't be supported in future
+ * @see <a href="https://spring.io/projects/spring-security-oauth">Spring
+ *      Security OAuth2 deprecation notice</a>
+ * @see <a href=
+ *      "https://github.com/spring-projects/spring-security/wiki/OAuth-2.0-Migration-Guide">OAuth
+ *      2.0 Migration Guide </a>
  */
 @Deprecated
 public class OAuth2AuthenticationConverter extends TokenAuthenticationConverter {
@@ -29,23 +30,7 @@ public class OAuth2AuthenticationConverter extends TokenAuthenticationConverter 
 		OAuth2AccessToken accessToken = new OAuth2AccessToken(OAuth2AccessToken.TokenType.BEARER, jwt.getTokenValue(),
 				jwt.getIssuedAt(), jwt.getExpiresAt());
 
-		OAuth2AuthenticatedPrincipal authenticatedPrincipal = new OAuth2AuthenticatedPrincipal() {
-			@Override
-			public Map<String, Object> getAttributes() {
-				return authenticationToken.getTokenAttributes();
-			}
-
-			@Override
-			public Collection<? extends GrantedAuthority> getAuthorities() {
-				return authenticationToken.getAuthorities();
-			}
-
-			@Override
-			public String getName() {
-				return authenticationToken.getName();
-			}
-		};
-
-		return new BearerTokenAuthentication(authenticatedPrincipal, accessToken, authenticationToken.getAuthorities());
+		return new BearerTokenAuthentication(new OAuth2Principal(authenticationToken), accessToken,
+				authenticationToken.getAuthorities());
 	}
 }